disconnect FilterNav and turn it into a component
