+++ src/containers/Explorer/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { ExplorerBody, FilterNav } from 'containers';
+import { ExplorerBody } from 'containers';
+import { ExplorerFilterNav } from 'containers/FilterNav';
 import { FilterNavManager } from 'hocs';
 import { withRouter } from 'react-router-dom';
 import { compose } from 'redux';
@@ -60,7 +61,7 @@ class Explorer extends React.Component {
     return (
       <div className={`${styles.explorerContainer}`}>
         <div className={styles.desktopFilter}>
-          <FilterNav
+          <ExplorerFilterNav
             defaultPlatforms={[config.postingPlatform]}
             position="fixed"
           />
